Move query and sync screens
