+++ util/src/org/commcare/util/screen/SyncScreen.java
@@ -1,25 +1,9 @@
-package org.commcare.util.cli;
+package org.commcare.util.screen;
 
 import org.commcare.modern.session.SessionWrapper;
-import org.commcare.modern.util.Pair;
 import org.commcare.session.CommCareSession;
-import org.commcare.session.RemoteQuerySessionManager;
-import org.commcare.suite.model.DisplayUnit;
-import org.javarosa.core.model.instance.ExternalDataInstance;
-import org.javarosa.core.model.instance.TreeElement;
-import org.javarosa.xml.ElementParser;
-import org.javarosa.xml.TreeElementParser;
-import org.javarosa.xml.util.InvalidStructureException;
-import org.javarosa.xml.util.UnfullfilledRequirementsException;
-import org.kxml2.io.KXmlParser;
-import org.xmlpull.v1.XmlPullParserException;
 
-import java.io.IOException;
-import java.io.InputStream;
 import java.io.PrintStream;
-import java.net.URL;
-import java.util.Hashtable;
-import java.util.Map;
 
 /**
  * Screen that displays user configurable entry texts and makes
